finsihed the social sections functionality

+++ lib/modals/view_model.dart
@@ -3,17 +3,16 @@ class ViewModel {
   int? contacted;
   double? tapThroughRate;
   String? lastContactDate;
-  ViewModel({
-    required this.views,
-    required this.contacted,
-    required this.tapThroughRate,
-    required this.lastContactDate
-});
+  ViewModel(
+      {required this.views,
+      required this.contacted,
+      required this.tapThroughRate,
+      required this.lastContactDate});
 
   factory ViewModel.fromJson(Map<String, dynamic> json) => ViewModel(
-    views: json['views'],
-    contacted: json['contacted'],
-    lastContactDate: json['lastContactDate'],
-    tapThroughRate: json['tapThroughRate'],
-  );
+        views: json['views'],
+        contacted: json['contacted'],
+        lastContactDate: json['lastContactDate'],
+        tapThroughRate: json['tapThroughRate'].toDouble(),
+      );
 }
